+++ lib/vagrant/action/builtin/box_remove.rb
@@ -60,18 +60,16 @@ module Vagrant
                 provider: box_provider.to_s,
                 versions: all_versions.join(", ")
             end
-          end
-
-          box = env[:box_collection].find(
-            box_name, box_provider, box_version)
-
-          if box == nil
+          elsif !all_versions.include?(box_version)
             raise Errors::BoxRemoveVersionNotFound,
               name: box_name,
               provider: box_provider.to_s,
               version: box_version,
               versions: all_versions.join(", ")
           end
+
+          box = env[:box_collection].find(
+            box_name, box_provider, box_version)
 
           env[:ui].info(I18n.t("vagrant.commands.box.removing",
                               :name => box.name,